add parametrize in test_grnrrators
